legacy-carwash/CarwashRavshans/carwash_project/requirements.txt
fastapi

uvicorn[standard]
sqlalchemy

pydantic[email]
pyyaml

python-jose[cryptography]
passlib[bcrypt]

websockets
python-multipart

aiosqlite

# Payment gateways
requests
httpx

# Additional utilities
schedule

# Excel export
openpyxl

# --- Testing ---
pytest
httpx
pytest-asyncio